Test: `LottoMachine` 테스트 분리
