+++ __tests__/lotto.test.js
@@ -1,23 +1,8 @@
 import Lotto from '../src/js/domain/Lotto.js';
 import LottoMachine from '../src/js/domain/LottoMachine.js';
-import {
-  DEFAULT_LOTTO_NUMBERS,
-  LOTTO_REWARD_DUMMY,
-  MATCHED_BONUS,
-  UNMATCHED_BONUS,
-} from './constants/lotto.js';
+import { DEFAULT_LOTTO_NUMBERS, LOTTO_REWARD_DUMMY, MATCHED_BONUS, UNMATCHED_BONUS } from './constants/lotto.js';
 
 describe('로또 구매 테스트', () => {
-  it.each([
-    [1_000, 1],
-    [10_000, 10],
-  ])('금액을 입력하면 장당 가격에 비례한 로또를 발급받는다.', (money, sheets) => {
-    const lottoMachine = new LottoMachine();
-    const lottos = lottoMachine.buy(money);
-
-    expect(lottos.length).toBe(sheets);
-  });
-
   it('로또는 여섯개의 숫자로 이루어진 배열을 매개변수로 받으며 그 값을 필드에 소유한다.', () => {
     const lottoMachine = new LottoMachine();
     const [lotto] = lottoMachine.buy(1000);
@@ -34,15 +19,12 @@ describe('로또 결과 확인 테스트', () => {
     [LOTTO_REWARD_DUMMY.FOURTH, 4],
     [LOTTO_REWARD_DUMMY.FIFTH, 3],
     [LOTTO_REWARD_DUMMY.LOST, 0],
-  ])(
-    '로또는 입력받은 배열과 비교하여 몇개가 동일한지 알 수 있다.',
-    (winningNumbers, matchCount) => {
-      const lotto = new Lotto(DEFAULT_LOTTO_NUMBERS);
-      lotto.check(winningNumbers, MATCHED_BONUS);
+  ])('로또는 입력받은 배열과 비교하여 몇개가 동일한지 알 수 있다.', (winningNumbers, matchCount) => {
+    const lotto = new Lotto(DEFAULT_LOTTO_NUMBERS);
+    lotto.check(winningNumbers, MATCHED_BONUS);
 
-      expect(lotto.matchCount).toBe(matchCount);
-    }
-  );
+    expect(lotto.matchCount).toBe(matchCount);
+  });
 
   it('로또는 보너스 번호가 자신의 번호에 포함되는지 알 수 있다.', () => {
     const lotto = new Lotto(DEFAULT_LOTTO_NUMBERS);
